# Optimize `IOSource#read_until` method