--- a/test/test_document.rb
+++ b/test/test_document.rb
@@ -404,7 +404,7 @@ EOX
         end
       end
 
-      class Read_until_Test < Test::Unit::TestCase
+      class ReadUntilTest < Test::Unit::TestCase
        def test_utf_8
           xml = <<-EOX.force_encoding("ASCII-8BIT")
 <?xml version="1.0" encoding="UTF-8"?>