@naitoh 2024-10-09
- Fixed to work for `UTF-16`.
- Added a test case for `UTF-16`.

- Does REXML support UTF-32?
    - I created #212.